Crate mem4

mem4 game - Learning Rust and Wasm/WebAssembly with Virtual Dom Dodrio and WebSocket communication
mem4 is a simple memory game made primarily for learning the Rust programming language.

Virtual DOM

Constructing a HTML page with Virtual DOM (vdom) is simple because it is rendered completely on every tick (animation frame).
Sometimes is hard for the developer to think what should change in the UI when some data changes.
The data can change from many different events and very chaotically (asyncronously).
It is easier to think how to render the complete DOM for the given data.
The dodrio library has ticks, time intervals when it do something.
If a rendering is scheduled, it will be done on the next tick.
If a rendering is not scheduled I believe nothing happens.
This enables asyncronous changing of data and rendering. They cannot happen in the same moment.
When GameData change and we now it will affect the DOM, then rendering must be scheduled.

State - action - message - state

In one moment the game is in a cetrain GameState. The user then makes an action. This action changes the GameData.
Then a message is sent to other players so they can also chenge their local GameData.
The rendering is scheduled and it will happen eventually. The result is a new GameState.

Futures and promises, Rust and JavaScript

Javascript is all asyncronous. Wasm is nothing else then a shortcut to the javascript engine.
So everything is asynchronous. This is pretty hard to grasp. Everything is Promises and futures.
Still better then CallBacks, but very hard to understand. And there is a constant jumping from thinking in Rust to thinking is JavaScript and back that is pretty confusing.
Javascript does not have a good idea of Rust datatypes. All there is is a JSValue type.
The library wasm-bindgen has made a fantastic job of giving to Rust the ability to call anything JavaScript can call, but the wayof doing it is sometimes very hard to understand.

Browser console

At least in modern browsers (FireFox and Chrome) we have the developer tools F12 and ther is a console we can output to. So we can debug what is going on with our program.

Modules

cardmoniker
cardmoniker.rs - renders the card moniker (card name/title)

gamedata
game data - structs and methods around game data

gridcontainer
gridcontainer.rs - renders the grid container with the images and most important the onclick event

playeractions
playeractions - renders the div to inform player what to do next and get a click action from the user

playersandscores
playersandscores.rs - renders the div that shows players and scores

rootrenderingcomponent
rootrenderingcomponent.rs - renders the web page

rulesanddescription
rulesanddescription.rs - renders the div that shows rules and descriptions All is a static content. Great for implementing dodrio cache.

websocketcommunication
websocketcommunication.rs - module that cares about websocket communication

Functions

run
To start the Wasm application, wasm_bindgen runs this functions

session_storage
Get the top-level window's session storage. TODO: to save user preferences maybe?